Finish Registration, improve safety
We can't rely on there being a response present if we have an error
during registration, so unfortunately we can't show the received error
code and the URL which we were attempting to contact. Also, "enabled"
doesn't appear to be present on the service, so we can just use the
number of enterprise hosts present. If none are configured, Enterprise
is disabled. You can comment things out in TOML if you want to
temporarily disable Enterprise.