Add enterprise admin interface
In order to interact with kapacitor from enterprise in a privileged and
authenticated manner, we have an admin API that is shipped along with
the enterprise-client. The ports on which this runs is configureable by
the end user in the stats config. Because this API needs to perform
request signing, it needs to wait for the token and secret from
enterprise, which is determined during registration with enterprise.