Factor out the protobuf dependency into a shell script.
The kapacitor unit tests don't run locally unless the protobuf
dependencies are installed locally.

Since the required glue is somewhat non-trivial, factor the glue
out from the Dockerfiles and change the docker files to use the same
glue.

Add a comment to the test to indicate how to fix the missing dependencies
issue.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>