Add Templates to Kapacitor.
TICKscript now allows vars to be overwritten via the API on a per task
basis.
TICKscript vars can now be any type and can be the result of an
expression involving other vars.

A template represents a TICKscript and task type that can be shared
across many tasks.
Each task can supply its own vars to the template.