modifying bufpool to provide closingBuffer
bufpool provided buffers result in race condition when used with a
buffer user that expects to take ownership of the buffer (like http
client appears to).  Returning an extended version of Buffer that allows
it to be closed, where the closing the buffer performs a Reset() on the
buffer and returns it to the pool it came from.