fix: return an error when aggregating a point fails
When the `AggregatePoint` method failed to convert the point to the
proper type, an error would happen and the underlying influxql code
would never receive the point within the aggregator.

But, the batch size would be incremented because the error was swallowed
instead of being reported back to the task node. This meant the task
believed that at least one point had been aggregated and it would
attempt to emit a point from aggregators that could not emit without at
least one point.

This change fixes the task so that it reports the error when it happens
instead of swallowing the error. This way, the error doesn't hide itself
with little evidence.

This has also been changed so the batch size is only incremented when an
error doesn't happen. The code had a path flow of incrementing the batch
size before it knew if an error happened when aggregating or not.