Update RPM specific packaging to match RPM naming convention
fix: Update RPM specific packaging to match RPM naming convention

RPM naming convention says that:
 1. We should use dashes not underscores in the package filename
 2. We should include a package version (iteration) in the package name

This change removes the code which swapped underscores for dashes in the RPM specific code. 
It also moves the code which removes the package_iteration in to an else block and adds an rpm specific work around to not do that.  This was done so that non-rpm builds are not affected by this change.

Fixes: #95